Separated Preprocessing and Path Finding Files for easy readability and fragmentation of code.
Preprocessing module contains the functions for preprocessing, image
display, finding start point and end point of the skeleton array.
This module was earlier contained in the trial1.py module itself.
Trial1.py has been modified with removed code, and incompletely
implemented code written for pruning and path finding.
The code is just trial based, and would be completed in some time now.
A documentation is also in the pipes and would be written and published
soon.